+++ plugins/woocommerce-admin/packages/admin-e2e-tests/src/pages/WcHomescreen.ts
@@ -1,21 +1,28 @@
 /**
+ * External dependencies
+ */
+import { ElementHandle } from 'puppeteer';
+
+/**
  * Internal dependencies
  */
 import {
 	waitForElementByText,
+	getElementByAttributeAndValue,
 	waitForElementByTextWithoutThrow,
+	getElementByText,
 } from '../utils/actions';
 import { BasePage } from './BasePage';
 
 export class WcHomescreen extends BasePage {
 	url = 'wp-admin/admin.php?page=wc-admin';
 
-	async isDisplayed() {
+	async isDisplayed(): Promise< void > {
 		// Wait for Benefits section to appear
 		await waitForElementByText( 'h1', 'Home' );
 	}
 
-	async possiblyDismissWelcomeModal() {
+	async possiblyDismissWelcomeModal(): Promise< void > {
 		const modalText = 'Welcome to your WooCommerce store’s online HQ!';
 		const modal = await waitForElementByTextWithoutThrow(
 			'h2',
@@ -33,7 +40,7 @@ export class WcHomescreen extends BasePage {
 		}
 	}
 
-	async getTaskList() {
+	async getTaskList(): Promise< Array< string | null > > {
 		await page.waitForSelector(
 			'.woocommerce-task-card .woocommerce-task-list__item-title'
 		);
@@ -51,7 +58,14 @@ export class WcHomescreen extends BasePage {
 		} );
 	}
 
-	async clickOnTaskList( taskTitle: string ) {
+	async isTaskListDisplayed(): Promise< boolean > {
+		return !! ( await waitForElementByTextWithoutThrow(
+			'*',
+			'Get ready to start selling'
+		) );
+	}
+
+	async clickOnTaskList( taskTitle: string ): Promise< void > {
 		const item = await waitForElementByText( '*', taskTitle );
 
 		if ( ! item ) {
@@ -64,10 +78,61 @@ export class WcHomescreen extends BasePage {
 		}
 	}
 
-	async waitForNotesRequestToBeLoaded() {
-		return await this.page.waitForResponse( ( response ) => {
+	async hideTaskList(): Promise< void > {
+		const taskListOptions = await getElementByAttributeAndValue(
+			'button',
+			'title',
+			'Task List Options'
+		);
+		await taskListOptions?.click();
+		await waitForElementByText( 'button', 'Hide this' );
+		const hideThisButton = await getElementByText( 'button', 'Hide this' );
+		await hideThisButton?.click();
+		await this.page.waitFor( 500 );
+	}
+
+	async waitForNotesRequestToBeLoaded(): Promise< void > {
+		await this.page.waitForResponse( ( response ) => {
 			const url = encodeURIComponent( response.url() );
 			return url.includes( '/wc-analytics/admin/notes' ) && response.ok();
 		} );
+	}
+
+	async isActivityPanelShown(): Promise< boolean > {
+		return !! ( await this.page.$( '.woocommerce-activity-panel' ) );
+	}
+
+	async getActivityPanels(): Promise<
+		Array< { title: string; count?: number; element?: ElementHandle } >
+	> {
+		const panelContainer = await page.waitForSelector(
+			'.woocommerce-activity-panel'
+		);
+		const list = await panelContainer.$$( 'h2' );
+		return Promise.all(
+			list.map( async ( item: ElementHandle ) => {
+				const textContent = await page.evaluate(
+					( el ) => el.textContent,
+					item
+				);
+				const match = textContent?.match( /([a-zA-Z]+)([0-9]+)/ );
+				if ( match && match.length > 2 ) {
+					return {
+						title: match[ 1 ],
+						count: parseInt( match[ 2 ], 10 ),
+						element: item,
+					};
+				}
+				return { title: textContent };
+			} )
+		);
+	}
+
+	async expandActivityPanel( title: string ): Promise< void > {
+		const activityPanels = await this.getActivityPanels();
+		const panel = activityPanels.find( ( p ) => p.title === title );
+		if ( panel ) {
+			await panel.element?.click();
+		}
 	}
 }