+++ plugins/woocommerce-admin/packages/admin-e2e-tests/src/specs/index.ts
@@ -4,3 +4,5 @@ export * from './analytics/analytics';
 export * from './analytics/analytics-overview';
 export * from './marketing/coupons';
 export * from './tasks/payment';
+export * from './homescreen/task-list';
+export * from './homescreen/activity-panel';