+++ plugins/woocommerce-admin/packages/admin-e2e-tests/typings/index.d.ts
@@ -1,1 +1,2 @@
 declare module '@woocommerce/e2e-environment';
+declare module '@woocommerce/e2e-utils';