+++ plugins/woocommerce-admin/tests/e2e/docker/initialize.sh
@@ -11,3 +11,6 @@ wp plugin install woocommerce --activate
 wp theme install twentynineteen --activate
 wp plugin activate woocommerce-admin
 wp user create customer customer@woocommercecoree2etestsuite.com --user_pass=password --role=customer --path=/var/www/html
+
+# we cannot create API keys for the API, so we using basic auth, this plugin allows that.
+wp plugin install https://github.com/WP-API/Basic-Auth/archive/master.zip --activate