+++ plugins/woocommerce-admin/tests/e2e/specs/homescreen/activity-panel.test.ts
@@ -1,0 +1,3 @@
+const { testAdminHomescreenActivityPanel } = require( '@woocommerce/admin-e2e-tests' );
+
+testAdminHomescreenActivityPanel();