+++ plugins/woocommerce-admin/packages/admin-e2e-tests/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+-   Add test cases for the home screen tasklist and activity panels. #7509
+
 # 0.1.0
 
 -   Released initial package